# Refactor CartesianParameterIterator: Depend on CartesianIterator