--- a/composer.json
+++ b/composer.json
@@ -21,6 +21,7 @@
         "beberlei/assert": "^2.4|^3.0",
         "doctrine/annotations": "^1.2.7",
         "lstrojny/functional-php": "1.0|^1.2.3",
+        "patchranger/cartesian-iterator": "~0.0.5",
         "phpbench/container": "~1.2",
         "phpbench/dom": "~0.2.0",
         "seld/jsonlint": "^1.0",