Didn't test it thoroughly, did just a few smoke-tests, looks good to me.
The idea is to get rid of self-crafted implementation of Cartesian product, relying on "mine-crafted" (read as "crafted by me") version of CartesianIterator, which in turn relies on standard, built-in [`MultipleIterator`](http://php.net/manual/en/class.multipleiterator.php), look how laconic it is: https://github.com/PatchRanger/cartesian-iterator/blob/master/src/CartesianIterator.php .
Could bring performance boost - but not for sure.