@PatchRanger 2018-11-17
I haven't fully debugged yet: some tests still fail.
I decided to run a benchmark to see whether my proposal is better in terms of efficiency. Here are the results:

Original

benchmark | subject | groups | params | revs | its | mem_peak | best | mean | mode | worst | stdev | rstdev | diff
-|-|-|-|-|-|-|-|-|-|-|-|-|-
RunBench | benchRun | | [] | 1 | 10 | 5,483,640b | 0.307s | 0.312s | 0.314s | 0.316s | 0.003s | 1.03% | 1.00x
RunBench | benchRunAndReport | | [] | 1 | 10 | 5,838,805b | 0.311s | 0.317s | 0.318s | 0.321s | 0.004s | 1.14% | 1.02x

PR

benchmark | subject           | groups | params | revs | its | mem_peak   | best   | mean   | mode   | worst  | stdev  | rstdev | diff
-|-|-|-|-|-|-|-|-|-|-|-|-|-
RunBench  | benchRun          |        | [[]]   | 1    | 10  | 5,490,450b | 0.324s | 0.427s | 0.369s | 0.789s | 0.136s | 31.92% | 1.00x
RunBench  | benchRunAndReport |        | [[]]   | 1    | 10  | 5,845,747b | 0.339s | 0.431s | 0.373s | 0.651s | 0.102s | 23.67% | 1.01x

It definitely looks like my implementation is less effective.
I guess I should close the PR.