@dantleech 2018-11-17
Thanks for the effort, it would indeed be nice to replace some of PHPBench's internals with better implementations, but would rather do this with established libraries, for example https://github.com/markrogoyski/math-php - but if it ain't broke, don't fix it :)